requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9